pyproject.toml
[tool.poetry]
name = "web-homework"
version = "0.1.0"
description = ""
authors = ["Mikhail Belin <belinma@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.13.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
